Gallery+Generate: Allow Recall of image params from the Gallery to Generate
This is a handy feature that allows users to find an image in their gallery that they liked and might want to further iterate on.

Upon clicking the "Recall" button in the Gallery, the user will be redirected to the Generate page, which will request the generation parameters from Navigator.

Those settings will be applied to the page (including the prompt, size, seed, model, etc).