Generate: After recalling & upscaling an image, reset the seed back to the random setting (-1).
This in order to recall & upscale an image, you need to have already generated the base image. So effectively, keeping the same seed after the upscale finishes doesn't make sense as if you click "Generate" without changing the seed, you'll just end up with a duplicate copy of the base image that you started with.

While yes, one could argue that you wanted to try the base image again with other settings (such as the Sampler/CFG Scale), in my personal experience this isn't super likely.

Nothing stops you from just recalling the image again to have the seed set back to the image's seed, it just means that you now have to explicitly do this.