Generate: Disable editing of prompt and other settings when an image is currently recalled
If a user recalls an image, they are either trying to replicate the image, or they are wanting to make variations. In both cases, editing the prompt (or other parameters) would invalidate the resulting image.

Since during recall-based operations we force set the prompt (and all other parameters) to that of the recalled image, then we might as well disable it. This way, the user doesn't have any false beliefs that their changes will actually apply on the resulting image (and will just be undone in the UI).