Generate: Hide post-actions "Recall Parameters" button
It's not really all that useful, because if you've just finished generating an image, likely all the parameters that were used to create it are already set. In reality, what ends up happening is you go to click delete or subtle variation and accidentally click it instead.

Of course, for the few situations where you *would* need to do this, you can just go back to the gallery page and trigger a recall from there.