Generate: Add in a field to recall an image from a given Job ID.
Technically, you can already do this by using the recall query parameter on the generate route... However, this isn't very user-friendly if someone has an image that they want to share the generation parameters of with a friend or such.

The field is within the "Generation Settings" menu, so is hidden when that section is collapsed, and additionally it will be hidden if there is already a recalled image active.