Settings: Refactor settings view to dynamically load components
Replaced hardcoded component imports with dynamic imports. This improves readability a bit, as later on if we have 10 settings categories, then we don't need to handwrite in the DOM element of each one. It also reduces initial load time by loading components on demand when a category is selected.

In theory, the trade-off would be that there could be a lag time when clicking on a category due to the browser needing to fetch the component - however, the components are ~1kb and are cacheable so it should be fine.