Settings: Add image generation settings panel with upscaling options
Introduced a new settings component for managing image generation configurations, including upscale steps, similarity, and upscaler selection. Updated relevant styles, settings store, and navigation to support the new options, which should help support power users in customizing image generation behaviors.

One core value that I am trying to keep with the design of the setting options is the first option is always the "default", so that way users do not need to recall what the original setting is should they wish to go back to the defaults.

Since some of these options can be changed rapidly (such as upscale similarity), we ensure that all settings get saved instantly and the user is told of this on the new settings page - this way, we do not need to provide feedback (which could get overwhelming) every time each single setting option is incremented.

Finally, another core value I am trying to keep is that there should be an explanation for all settings (only exception should be possible settings that really are just "common sense") - so that even if the user isn't a power-user but still wants to learn more, they can do so!

It might not be possible to explain *everything*, but it gives users a jump-starting point so that if they wish to research it further, they should hopefully know what to look for.