Generate: Rework Recall logic where possible
Attempt to avoid doing a recall if we really don't need it. Recall API requests are heavy (as the backend needs to re-read the image metadata to return it), which can delay UI responses.

Additionally, since we lock most UI settings (for good reason), it should be used as sparingly as possible.

Initial tests didn't show any problems with this approach, but we might need to revert this later if it turns out to cause issues.