Generate: Ignore rescale meta on recall
When an image is recalled, there's not really much of a use case for setting the UI settings to target upscale size values if present.

Might look into moving this to a setting later on if a use case can be found, but for now only interpret the original size values of the image's metadata.