Application: Update Alerts to use Toasts
The banner alerts are fine for persistent alerts (such as indicating setting sync status)... but for temporary alerts the constant show/hide of the element causes other page elements to shift which is annoying UX wise.

Now we utilize Toasts which will appear at the top of the page. On smaller displays, it will span the entire width, and on large displays it will be the size of the text itself and anchor to the top left.

Previously, the Predefined Prompts settings area used Toasts with its own Toast system. Since this is now universal across the application, refactored the Predefined Prompts area to use the application-wide alert system.