LoRA+Generate: Update LoRA selections when recalling an image
This was far more complex than I expected. Detecting the LoRAs that were enabled from the prompt was fairly easy. Figuring out how to "pass down" the fact that an image was recalled (so that the LoRA components can recompute their states) was also relatively easy.

Figuring out how to *remove* the LoRA activation sequence(s) from the prompt whenever an image was recalled (as otherwise it would be duplicated when sending a new request if not manually cleared) was a royal pain.

It looks like there are still some edge cases where the activation sequences remain in the prompt (somewhere along the process of setting the strengths for activated LoRAs causes a shadow entry to get added in possibly?) but these seem to be fairly rare, hopefully (and will be worked on later).