Generate+LoRA: Re-allow editing of LoRA settings while a job is running
Since all other fields can continue to be modified while a generation request is running, and that changing LoRA settings does not launch a concurrent generation request it does not make much sense to have it disabled.

It will however continue to stay disabled when a recalled image is active, for the same reason that all other fields are disabled (so that variations still work correctly).