Gallery+Generate+Settings: Add share to CivitAI buttons
To see these buttons, they must be enabled (defaults to disabled) in the `Settings -> Image Generation` menu (probably needs a better category to go under later). When enabled, the buttons will be displayed in the gallery and on the post-processing section of the generation page.

The button uses CivitAI's Post Intent feature; when clicking it opens a new tab to CivitAI with the image URL encoded as the `mediaU` parameter.

Unfortunately, because this is a "pull" from CivitAI's side rather than a "push" from Compass/Navigator this means that in order for it to work, the Navigator instance that Compass is connected to will need to be public (or at least, the image viewing endpoint) as otherwise CivitAI won't be able to pull the image.