Feature: Add Model Browser/Installer feature with CivitAI integration
Introduced a Model Browser that integrates with the CivitAI API, allowing users to search, filter, and download models. Added required components, services, and route updates, with access restricted to users with role 2 ("Artificer") or higher.

API requests are proxied through Navigator to avoid CORS issues as well as ensuring that a valid API key is attached. Performs an API request against Navigator first to ensure that the feature can be used.

If the API indicates that it is not available, then the navigation entries will be hidden. The user can still navigate to the page manually and browse models (such as to grab a link to provide to the operator) - but the download button is hidden and the user will be warned that they cannot trigger downloads via the UI.