Navigation: Don't check downloads enabled status if not logged in
Otherwise, this causes a bit of a race condition where the navigation bar will trigger an API request that requires authentication before the authentication state is synced. Since this throws a 401, the application thinks the token has expired and logs the user out. It also causes a session-expired warning when logging in due to the same issue.

So for now, just don't do this API request if the user isn't logged in. Unfortunately, it means that when the user has logged in (a fresh login), the model browser entry will still be hidden since the check was done before the login completed. Just refreshing the page will fix this, though.

To solve this long term, we'll want to look into having an application-wide event bus, and when authentication occurs, redo this check.