Home: Update home design; Add a carousel of example SD images
The home page now looks significantly better than before, it uses a carousel of pre-generated example stable diffusion images that are nice, high-resolution, and should be safe for public display.

Operators of Compass can replace the Carousel images by placing images into the `./public/example-images` folder and running the script from `./scripts/create_manifest.sh`.

The script requires `jq`, but if it is not available, the JSON manifest is simply an array of file names in the `./public/example-images` folder.