Generate: Refactor component structure; Misc changes/features; Disable Img2Img
Updated the grid layout of this component so that the "Generation Settings" being expanded no longer causes the "Post Processing" section to move down with it.

Added a few UI glow effects to the "Post Processing" area along with some buttons on the page.

Disabled the ability to navigate to the Img2Img tab from the UI, as the API seems broken currently. The tab selector will appear if the user accidentally directly navigates to the URL (as that was kept in place for easier debugging) just so that they can switch back.

Updated the "Results" image display to scale better for varying image sizes. Some images may require expanding into full screen to view full detail - this can be done by just clicking the image. This brings up a modal with a fuller preview of the image, can be dismissed by clicking anywhere or pressing Escape on the keyboard.

Updated the border colors of the various UI controls to better match the new base application colors.

Reduced default steps and cfg_scale for better performance (quicker, and adheres better to general prompts).

The prompt area received multiple UI updates; Positive/Negative prompts have different text color hints (green for positive - red for negative). Adjusted the text sizing of both prompt areas, along with the placeholder text.

The prompt text areas will also resize according to how much text is entered in them - updates upon text change.

Further updated the logic to disable components quicker when sending off generation requests - try to less rely on Navigator confirming the request so that the UI feels more reactive to these.

Try to ensure the progress bar and progress text label has some sort of indication before Navigator acknowledges the generation request for the same reasons as above (ensure the UI feels more reactive).

When a seed is set in image parameters, add a "Generate Subtle/Strong Variation" buttons to the main generation controls. Also change the text of the "Generate" button to "Regenerate From Seed" just so that the user is more aware that their settings will in a specific generation.

Updated the "Note: You have a seed set..." label to be clickable to re-randomize the seed; Update said label's text to indicate this. This reduces the amount of navigation the user has to do to re-randomize (no longer need to expand generation settings, scroll to advanced options, etc).