Home: Switch images to WebP format and update carousel logic
Replaced PNG images with optimized WebP versions in `image-manifest.json` and the `example-images` directory to improve performance. It does result in the images being slightly lower quality, but it should be fine.

Updated the carousel logic to shuffle and display only the first five images to loading overhead.

(There are still enough images in total to choose from so that it still hopefully looks "unique" during a revisit/reload of the page)