CI: Add GitHub workflow and Docker setup for Compass builds
Introduce a GitHub Actions workflow to build and push Compass images to GHCR on main and dev branch updates.

Add a Docker image configuration (docker/compass-prebuilt) with Caddy server (same as the original) and runtime environment variable configuration so that these variables can still be configured post-build.

Users of this image are expected to pass the `COMPASS_APP_` variables to their container so that the entrypoint script can replace the variable references with their correct contents.