Model Browser: Allow searching by Model ID
Implementation is definitely "hacky", but CivitAi's regular model search
API is not handling name searches as well as it used to at the moment.

Searching by ID should have a dedicated UI element instead of "If query
is fully numeric, assume model ID" but this is what we'll use for now.