LoRA+Generate: Refactor LoRA selector to LoRA Browser
This migrates the old "LoRA Selector" to a browser instead. The old
selector was just a plain dropdown component which is fine for a small
handful of LoRAs, but once you start getting past 25 or so LoRAs it
becomes difficult to browser a dropdown.

The new browser allows for searching by name, filtering by base model
type, shows training words (though this may be removed due to the
varying length of these lists and its not super important to the user
until its activated - we'll see), and will eventually show an example
image perhaps.

Currently, Navigator discards the metadata that would contain the
example images so we need to correct that in Navigator first before we
can implement those here.